feat: graph-based computation for witness commitment generation. (#45)
This PR brings a first shot to transforming the prover code into a graph
of executable tasks, just by starting with the commitment of the
witness. The endgoal is that the whole prover module is "graphized" such
that generating a proof means executing through the graph until
completion.
Currently, only a local executor is implemented inside zkml - the
"network" one can be done later in a separate module. This executor is
super simple: just execute all nodes "ready" until there is nothing else
to do.
I plan to actually add concurrency to the local executor with a thread
pool so _we could keep always all the cores busy_ : now that we express
computation as a graph, it becomes easy to always use all CPU available
by always run a task as soon as a thread gets freed.

---------

Co-authored-by: Franklin Delehelle <franklin.delehelle@odena.eu>