perf: implement qkv layer in burn (#55)
- I'm working separately on enabling burn fusion that is important for
this layer to perform better on gpu
(https://linear.app/lagrange/issue/DP-117/enable-burn-fusion-backend-for-gpu)
- Note that there are a lot of changed test inputs in regression tests
(job
https://github.com/Lagrange-Labs/deep-prove-private/actions/runs/17129492266/job/48589623418?pr=55).
This is due to a difference in our tensor's matmul vs. burn's. This
affects all the `q`, `k`, `v` fields of the layer (and layers are
included in regression test's hash of inputs).
- There's a cache field in `QKV` layer that is being de/se-rialized
together with the layer. Should we set this to `#[serde(skip)]`?
- I've checked that when I forced to fail all the tests that had `qkv`
layer included, so that no regression layer's outputs are captured for
those, the rest of the outputs were matching before and after as
expected