perf: enable burn fusion with gpu (#72)
- enabes "fusion" feature for burn that makes it fuse operations on
tensors into a single kernel (it significantly improves the gpu
performance of e.g. qkv layer #55)
- removes the custom gelu impl to use one provided by burn (I checked
that they produce same results with 1e-3 precision tolerance for both
cpu and gpu)

Note that burn is updated to an unreleased version because we need
https://github.com/tracel-ai/burn/pull/3575 (and
https://github.com/tracel-ai/cubecl/pull/831) needed for the activation
layer (and most likely more upcoming burn layer conversions). I tried to
backport this PR onto burn v0.18 but it was non-trivial so I think using
this version from main until these changes are released is safer

The failing regressions are expected due to floating point differences
in gelu/activation layer - I've checked that when I forced to fail all
the tests that have activation layer included, so that no regression
layer's outputs are captured for those, the rest of the outputs were
matching before and after as expected