feat: enable partition of a graph from coloring (#68)
The goal of this PR is lay the basis to partition the computational
graph according to the coloring.
The idea is that we only need to define the graph once, with a single
set of tasks, regardless if they happen locally on the worker or are
part of the exchange rounds between workers, and we can even apply
coloring afterwards if we want.
Next step after this is to define more portion of the prover as
`GraphNode` and the "split" will happen automatically then.

---------

Co-authored-by: Franklin Delehelle <franklin.delehelle@odena.eu>