refactor: make the graph module more generic (#83)
Before a `Graph` was specific for a node with a color and to be executed
within a `Scheduler`.
Now `Graph` is fully generic over the type of the node and the edge;
it's basically a smart wrapper around `petgraph::DiGraph` that gives out
some helper functions that are longer and trickier to use in petgraph.
The rest `Scheduler` and `Executor` all expects a `ColoredNode` as
usual, since it's for execution.

The goal after this PR is to re-use the `Graph` struct to replace the
homemade graph we have in `Model` and replace the
difficult-to-manipulate `Edge` structure. Given the amount of work
related to only graph manipulation that is starting due to the
chunking/graphization, this refactor is necessary to then cleanly
decouple the graph logic with the application using it , and makes us
having only one graph implementation, not two.