fix: quantised argmax now only considers unpadded dim size (#57)
The quantised version of Argmax was also searching for maximum values in
the padding part of the tensor. This PR fixes this by making the
verifier check the final output tokens are all within the model
vocabulary and additionally by using a less than polynomial in the
sumcheck step of proving.

---------

Co-authored-by: nicolas gailly <nikkolasg@gmail.com>