perf: changed profiling framework to divan (#105)
# Why?

criterion seems to always call the registered benchmark functions, and
only filter out the benchmark execution in the bencher methods, e.g.
`bench_with_input`. This was adding almost 40s of additional overhead to
running the benchmarks because the inputs are generated outside of these
function calls. moving the input generation to a callback like
[iter_batched](https://docs.rs/criterion/latest/criterion/struct.Bencher.html#method.iter_batched)
removes the need to generate inputs for benchmarks that are not
executed, but would also generate the inputs once per run instead of
once per benchmark `O(1)` -> `O(N)`.

the dilan libray doesn't have these issues, and seems to have a nicer
API. this also removes the configuration used by criterion that
drastically increases the execution time of fast tests.