feat: add GLU variant for activation (#116)
This PR modifies the activation layer to include a variant which also
multiplies an additional input tensor to the output of the activation
function. This variant is useful when the activation function is used as
part of a Gated Linear Unit (GLU).

Furthermore, the PR adds also the possibility to compute scaling factors
for intermediate tensors of a layer, which was necessary to properly
quantize the newly introduced variant of activation layer.

---------

Co-authored-by: Nicolas Gailly <nikkolasg@users.noreply.github.com>