refactor: homogeneize graph traversal (#173)
This PR implements a minimal (lol...) set of changes to reach a main
goal: make graph traversal homogeneous, which is particularly noticeable
_e.g._ in the more streamlined code for inference, proving, and
verifying, where we can get rid of off all the side-effects that had to
be handled to process model inputs and outputs.

Now, those are nodes like any other, which allow to get rid of
additional datasets to accumulate their own data (can't be more precise
as those depends of the function considered) and just treat them as a
particular case of a big `match` applied to every node.

I tried to keep the changes to the public API as few as can be, and
mostly only `Graph::set_{input, output}` got the axe.