feat: refactor parser logic and add safetensors format for Gemma3 (#168)
The idea is that defining a new format should be done in conjunction
with which model we wanna support.
It changes the paradigm from "let's try to guess what model and what
format you want to load" to end up closer to what other frameworks are
doing because there's no really other choices: you gotta know which
model you're loading and from what format.
That brings safetensors parsing as well now with Gemma3 (safetensors
because that's the easiest way to import models with the minimum loss of
accuracy AND to be able to accurately run traces of pytorch models when
it's using this format, vs GGUF that requires llama-cpp that doesn't
expose the capability to look for intermediate results).

---------

Signed-off-by: delehef <franklin.delehelle@odena.eu>
Co-authored-by: delehef <franklin.delehelle@odena.eu>
Co-authored-by: Zack Youell <zackyouell@mac.com>