feat: store interior mutability (#195)
Changes the store's trait to force interior mutability. This allows the
store to be copied and passed around freely, and it will still be
useful. Without the interior mutability only a single reference to the
store can be used.

As a side effect the `LocalStore` no longer implements the trait, it
become an implementation detail, otherwise that would require an
additional lock for interior mutability.