ci: bench on nix machines (#196)
Updates CI benches to run on our nix machines, which should hopefully
reduce noise from running on public machines.

The bencher binary used by the provided action doesn't work on nix
out-of-box (fails with `Could not start dynamically linked executable:
bencher`, because the `cargo-bencher` bin is just a thin wrapper over
`bencher_cli`), so we're building it from source instead (thx @delehef
for the idea). Note that we're installing `bencher_cli` with `--debug`
to speed up the action (30s instead of ~50min with release).
Optimizations for this bin don't matter as it's just a client talking to
bencher api server.