feat: chunking commitments (#198)
This is the final PR to implement the chunking of the prover, after the
first PR #170 that split the claims proving among multiple chunks.

- Split also the witness commitments and opening proofs among multiple
chunks
- Remove the `op` layer from the trace, fetching it from the set of the
nodes of the model, which now needs to be provided to the `prove`
methods. This allows to serialize the trace and reduce its size, given
that the trace is meant to be sent to prover nodes over the network
- Adds all the data structures to build an execution graph to prove
chunks, which allows to dispatch the computational tasks among a set of
workers. An example `chunk-prover` is added to showcase how the
execution graph for a model can be built and run using an `Executor`
implementation

---------

Co-authored-by: Nicolas Gailly <nikkolasg@users.noreply.github.com>